Suppress "method redefined" warning
This code emits the following warning to anyone who uses the gem,
unless they suppress it with something like the warning gem, or
suppress all warnings:

    lib/ruby/gems/3.2.0/gems/openapi_parser-2.1.0/lib/openapi_parser/schemas/schema.rb:113: warning: method redefined; discarding old additional_properties

This has been the case for about five years, although the comments
make reference to this situation being transitory.  It would be better
to use the gem without it emitting warnings in the meantime.