Remove unused codes (#162)
* Remove unused codes

Now that we support Ruby 2.7 and above, we can remove conditions with versions lower than that.

* Replace String#match with String#match?

`String#match?` is more suitable when only a boolean is needed, as it is faster than `String#match`. OpenapiParser now supports Ruby2.7 and above, and since `String#match?` has been available since Ruby2.4, it’s appropriate to make this switch.