Correctly validate 204 responses (#133)
* Do not raise error when body is empty and nil content type

Empty response bodies (204 responses for instance) should not have a
content type and therefore an error should not be raised in strict mode
when the response body is empty and the content type is nil as this
behavior is expected.

* Return true instead of nil when validation passes

As mentioned in #101 it makes more sense to return true from this method
when validation passes instead of nil.

---------

Co-authored-by: Tristan O'Neil <tristan.oneil@calendly.com>